requirements.txt
asgiref==3.4.1
attrs==21.2.0
base58==2.1.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.6
cheroot==8.5.2
CherryPy==18.6.0
construct==2.10.67
cryptography==3.4.7
Django==3.2.6
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.8.0
ed25519==1.5
idna==2.10
imgkit==1.2.2
importlib-metadata==4.8.1
iniconfig==1.1.1
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
more-itertools==8.8.0
packaging==20.9
pluggy==0.13.1
portend==2.7.1
psycopg2==2.7.7
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
solana==0.9.2
sqlparse==0.4.1
tempora==4.1.1
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uWSGI==2.0.19.1
websockets==9.1
zc.lockfile==2.0
zipp==3.5.0
